+++ dev-resources/user.clj
@@ -2,6 +2,7 @@
   (:require
     [criterium.core :as c]
     [clojure.java.io :as io]
+    com.walmartlabs.lacinia.expound
     [com.walmartlabs.lacinia.schema :as schema]
     [com.walmartlabs.lacinia :as l]
     [clojure.spec.alpha :as s]
@@ -10,3 +11,4 @@
 (require 'com.walmartlabs.lacinia.expound)
 
 (alter-var-root #'s/*explain-out* (constantly expound/printer))
+