+++ src/com/walmartlabs/lacinia/schema.clj
@@ -27,7 +27,8 @@
     [com.walmartlabs.lacinia.internal-utils
      :refer [map-vals map-kvs filter-vals deep-merge q
              is-internal-type-name? sequential-or-set? as-keyword
-             cond-let ->TaggedValue is-tagged-value? extract-value extract-type-tag]]
+             cond-let ->TaggedValue is-tagged-value? extract-value extract-type-tag
+             to-message]]
     [com.walmartlabs.lacinia.resolve :as resolve :refer [ResolverResult resolve-as combine-results is-resolver-result?]]
     [clojure.string :as str]
     [clojure.set :refer [difference]]
@@ -141,7 +142,11 @@
   The function is only invoked if the value to be conformed is non-nil.
 
   Any exception thrown by the function is silently caught and the returned conformer
-  will return :clojure.spec/invalid or a [[coercion-failure]]."
+  will return :clojure.spec/invalid or a [[coercion-failure]].
+
+  This function has been deprecated, as Scalar parse and serialize callbacks are now
+  simple functions, and not conformers."
+  {:deprecated "0.31.0"}
   [f]
   (s/conformer
     (fn [x]
@@ -153,46 +158,39 @@
             (coercion-failure message (ex-data e))
             ::s/invalid))))))
 
-(defn ^:private invalid-scalar
-  [type-name value]
-  (ex-info (format "Invalid %s value." (name type-name))
-           {:value (pr-str value)}))
-
-(defmacro ^:private catch-as-invalid
+(defmacro ^:private catch-as-coercion-failure
   [& body]
   `(try
      ~@body
      (catch Throwable _#
-       ::s/invalid)))
+       nil)))
 
-(defn ^:private parse-int
+(defn ^:private int-parse
   [v]
   (cond
+    ;; The serialized is a little more forgiving about converting non-integers to integers.
+    ;; On the parse side, we're a little more picky.
     (integer? v)
     (if (<= Integer/MIN_VALUE v Integer/MAX_VALUE)
       (int v)
-      (throw (ex-info "Int value outside of allowed 32 bit integer range." {:value v})))
+      (coercion-failure "Int value outside of allowed 32 bit integer range." {:value (pr-str v)}))
 
-    (string? v) (catch-as-invalid (Integer/parseInt v))
+    (string? v)
+    (catch-as-coercion-failure
+      (Integer/parseInt v))))
 
-    :else (throw (invalid-scalar :Int v))))
-
-(defn ^:private serialize-int
+(defn ^:private int-serialize
   [v]
   (cond
-    ;; Spec: should attempt to coerce raw values to int
     (string? v)
-    (let [v' (catch-as-invalid (Integer/parseInt v))]
-      (if (keyword? v')
-        v'
-        (recur v')))
+    (catch-as-coercion-failure (Integer/parseInt v))
 
-    (and (number? v)
-         (<= Integer/MIN_VALUE v Integer/MAX_VALUE))
-    (int v)
-
-    :else
-    (throw (ex-info "Int value outside of allowed 32 bit integer range." {:value (pr-str v)}))))
+    ;; Per the spec, if it's in range, we just truncate it to an int.
+    ;; Note that we should be rejecting floats that have a non-zero decimal amount.
+    (number? v)
+    (if (<= Integer/MIN_VALUE v Integer/MAX_VALUE)
+      (int v)
+      (coercion-failure "Int value outside of allowed 32 bit integer range." {:value (pr-str v)}))))
 
 (defn ^:private coerce-to-float
   [v]
@@ -205,17 +203,14 @@
     (double v)
 
     (string? v)
-    (catch-as-invalid (Double/parseDouble v))
-
-    :else
-    (throw (invalid-scalar :Float v))))
+    (catch-as-coercion-failure (Double/parseDouble v))))
 
 (defn ^:private string->boolean
   [^String s]
   (case s
     "true" true
     "false" false
-    (throw (ex-info "Boolean string must be `true' or `false'." {:value s}))))
+    (coercion-failure "Boolean string must be `true' or `false'." {:value s})))
 
 (defn ^:private parse-boolean
   [v]
@@ -224,38 +219,31 @@
     v
 
     (string? v)
-    (string->boolean v)
-
-    :else
-    (throw (invalid-scalar :Boolean v))))
+    (string->boolean v)))
 
 (defn ^:private serialize-boolean
   [v]
   (cond
-    (instance? Boolean v) v
+    (instance? Boolean v)
+    v
 
-    ;; Spec: coerce non-boolean raw values to Boolean when possible.
+    ;; coerce non-boolean raw values to Boolean when possible
     (number? v)
     (not (zero? v))
 
-    (string? v) (string->boolean v)
-
-    :else
-    (throw (invalid-scalar :Boolean v))))
+    (string? v) (string->boolean v)))
 
 (def default-scalar-transformers
-  (let [str-conformer (as-conformer str)
-        float-conformer (as-conformer coerce-to-float)]
-    {:String {:parse str-conformer
-              :serialize str-conformer}
-     :Float {:parse float-conformer
-             :serialize float-conformer}
-     :Int {:parse (as-conformer parse-int)
-           :serialize (as-conformer serialize-int)}
-     :Boolean {:parse (as-conformer parse-boolean)
-               :serialize (as-conformer serialize-boolean)}
-     :ID {:parse str-conformer
-          :serialize str-conformer}}))
+  {:String {:parse str
+            :serialize str}
+   :Float {:parse coerce-to-float
+           :serialize coerce-to-float}
+   :Int {:parse int-parse
+         :serialize int-serialize}
+   :Boolean {:parse parse-boolean
+             :serialize serialize-boolean}
+   :ID {:parse str
+        :serialize str}})
 
 (defn ^:private error
   ([message]
@@ -357,8 +345,14 @@
 (s/def ::input-object (s/keys :opt-un [::description
                                        ::directives]
                               :req-un [::fields]))
-(s/def ::parse s/spec?)
-(s/def ::serialize s/spec?)
+;; Prior to 0.31.0, specs were conformers.
+;; With the breaking change in 0.31.0, we want to make sure that custom scalars
+;; have been updated.
+(s/def ::not-a-conformer #(not (s/spec? %)))
+(s/def ::parse-or-serialize-fn (s/and ::not-a-conformer
+                                      fn?))
+(s/def ::parse ::parse-or-serialize-fn)
+(s/def ::serialize ::parse-or-serialize-fn)
 (s/def ::scalar (s/keys :opt-un [::description
                                  ::directives]
                         :req-un [::parse
@@ -642,7 +636,7 @@
         category (:category field-type)
 
         ;; Build up layers of checks and other logic and a series of chained selector functions.
-        ;; Normally, don't redefine local symbols, but here is makes it easier to follow and less
+        ;; Normally, don't redefine local symbols, but here it makes it easier to follow and less
         ;; brittle.
 
         selector floor-selector
@@ -650,20 +644,34 @@
         selector (if (= :scalar category)
                    (let [serializer (:serialize field-type)]
                      (fn select-coerion [selector-context]
-                       (let [{:keys [resolved-value]} selector-context
-                             serialized (s/conform serializer resolved-value)]
-                         (cond
+                       (cond-let
 
-                           (= serialized :clojure.spec.alpha/invalid)
-                           (selector-error selector-context (error "Invalid value for a scalar type."
-                                                                   {:type field-type-name
-                                                                    :value (pr-str resolved-value)}))
+                         :let [{:keys [resolved-value]} selector-context]
 
-                           (is-coercion-failure? serialized)
-                           (selector-error selector-context serialized)
+                         (nil? resolved-value)
+                         (selector selector-context)
 
-                           :else
-                           (selector (assoc selector-context :resolved-value serialized))))))
+                         :let [serialized (serializer resolved-value)]
+
+                         (nil? serialized)
+                         (selector-error selector-context
+                                         (let [value-str (pr-str resolved-value)]
+                                           {:message (format "Unable to serialize %s as type %s."
+                                                             value-str
+                                                             (q field-type-name))
+                                            :value value-str
+                                            :type-name field-type-name}))
+
+                         (is-coercion-failure? serialized)
+                         (selector-error selector-context
+                                         (-> serialized
+                                             (update :message
+                                                     #(str "Coercion error serializing value: " %))
+                                             (assoc :type-name field-type-name
+                                                    :value (pr-str resolved-value))))
+
+                         :else
+                         (selector (assoc selector-context :resolved-value serialized)))))
                    selector)
 
         selector (if (= :enum category)
@@ -1016,14 +1024,6 @@
            :values values
            :values-detail details
            :values-set values-set)))
-
-(defmethod compile-type :scalar
-  [scalar _]
-  (let [{:keys [parse serialize]} scalar]
-    (when-not (and parse serialize)
-      (throw (ex-info "Scalars must declare both :parse and :serialize functions."
-                      {:scalar scalar})))
-    scalar))
 
 (defmethod compile-type :object
   [object schema]