+++ test/com/walmartlabs/lacinia/validator_test.clj
@@ -277,7 +277,7 @@
                                      :value "hello world"}
                         :locations [{:column 3
                                      :line 1}]
-                        :message "Exception applying arguments to field `echoArgs': For argument `integer', scalar value is not parsable as type `Int'."}]}
+                        :message "Exception applying arguments to field `echoArgs': For argument `integer', unable to convert \"hello world\" to scalar type `Int'."}]}
              (execute compiled-schema q {} nil))))
 
   (testing "undefined argument"
@@ -299,15 +299,13 @@
                                       string: \"five\",
                                       nestedInputObject: {integerArray: \"hello world\",
                                                           date: \"1983-08-13\"}}) { integer }}"]
-      (is (= {:errors
-              [{:message
-                "Exception applying arguments to field `echoArgs': For argument `inputObject', scalar value is not parsable as type `Int'.",
-                :locations [{:line 1
-                             :column 3}]
-                :extensions {:field :echoArgs
-                             :argument :inputObject
-                             :value "hello world"
-                             :type-name :Int}}]}
+      (is (= {:errors [{:extensions {:argument :inputObject
+                                     :field :echoArgs
+                                     :type-name :Int
+                                     :value "hello world"}
+                        :locations [{:column 3
+                                     :line 1}]
+                        :message "Exception applying arguments to field `echoArgs': For argument `inputObject', unable to convert \"hello world\" to scalar type `Int'."}]}
 
              (execute compiled-schema q {} nil)))))
 
@@ -340,7 +338,8 @@
                                      :value "foo"}
                         :locations [{:column 2
                                      :line 1}]
-                        :message "Exception applying arguments to field `echoArgs': For argument `integerArray', scalar value is not parsable as type `Int'."}]}
+                        :message "Exception applying arguments to field `echoArgs': For argument `integerArray', unable to convert \"foo\" to scalar type `Int'."}]}
+
              (execute compiled-schema q {} nil)))))
 
   (testing "valid arguments"